[WebConsoleV2] Reorganize typings - defaults to use module
