+++ web_console_v2/client/src/services/user.ts
@@ -1,5 +1,6 @@
 import { AxiosPromise } from 'axios'
 import request from 'libs/request'
+import { FedLoginFormData, FedUserInfo } from 'typings/auth'
 
 export function fetchUserInfo(id: string): AxiosPromise<FedUserInfo> {
   return request(`/v2/auth/users/${id}`)
