--- web_console_v2/client/src/typings/component.d.ts
@@ -1,1 +1,0 @@
-declare type ComponentSize = 'small' | 'medium' | 'large' | 'default'
