+++ web_console_v2/client/src/typings/component.ts
@@ -1,0 +1,6 @@
+export type ComponentSize = 'small' | 'medium' | 'large' | 'default'
+
+export interface StyledComponetProps {
+  className?: string
+  [key: string]: any
+}
