+++ web_console_v2/client/src/typings/workflow.ts
@@ -1,3 +1,5 @@
+import { ComponentSize } from './component'
+
 export enum VariableComponent {
   Input = 'Input',
   Select = 'Select',
