+++ web_console_v2/client/src/views/routes.tsx
@@ -7,6 +7,7 @@ import Datasets from 'views/Datasets'
 import Settings from 'views/Settings'
 import CreateProject from 'views/CreateProject'
 import EditProject from 'views/EditProject'
+import { FedRouteConfig } from 'typings/app'
 
 const routes: FedRouteConfig[] = [
   {
@@ -14,56 +15,48 @@ const routes: FedRouteConfig[] = [
     exact: true,
     component: () => <Redirect to="/projects" />,
     auth: true,
-    children: [],
   },
   {
     path: '/dashboard',
     exact: false,
     component: Dashboard,
     auth: true,
-    children: [],
   },
   {
     path: '/settings',
     exact: false,
     component: Settings,
     auth: true,
-    children: [],
   },
   {
     path: '/projects',
     exact: true,
     component: Projects,
     auth: true,
-    children: [],
   },
   {
     path: '/workflows',
     exact: true,
     component: Workflows,
     auth: true,
-    children: [],
   },
   {
     path: '/datasets',
     exact: true,
     component: Datasets,
     auth: true,
-    children: [],
   },
   {
     path: '/create-project',
     exact: true,
     component: CreateProject,
     auth: true,
-    children: [],
   },
   {
     path: '/edit-project',
     exact: true,
     component: EditProject,
     auth: true,
-    children: [],
   },
 ]
